requirements.txt
pandas==1.5.3
numpy==1.23.4
tensorflow==2.11.0
matplotlib==3.7.1
opencv-python==4.7.0.72
Pillow==9.3.0
